requirements.txt
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.25.0
gymnasium>=0.29.0
numpy>=1.24.0
Pillow>=10.0.0
pyyaml>=6.0
wandb>=0.16.0
tqdm>=4.65.0
qwen-vl-utils
# LIBERO (install separately)
# pip install libero
